+++ Backend/src/main/java/com/pedro/cruzeiro/dev/inventorymanagement/dto/in/UpdateProductRequest.java
@@ -14,8 +14,6 @@ import java.util.List;
 @AllArgsConstructor
 public class UpdateProductRequest {
 
-  private String productId;
-
   private String name;
 
   private String description;