+++ Backend/src/main/java/com/pedro/cruzeiro/dev/inventorymanagement/exception/ProductUnitException.java
@@ -1,0 +1,8 @@
+package com.pedro.cruzeiro.dev.inventorymanagement.exception;
+
+public class ProductUnitException extends RuntimeException {
+
+  public ProductUnitException(String message) {
+    super(message);
+  }
+}